@@ -1,22 +1,12 @@
 <?php
 
-namespace Magezil\BoughtTogether\Helper;
+namespace Magezil\BoughtTogether\Model\Config\Source;
 
-use Magento\Framework\App\Helper\AbstractHelper;
 use Magento\Framework\App\Config\ScopeConfigInterface;
 use Magento\Store\Model\StoreManagerInterface;
 use Magento\Store\Model\ScopeInterface;
 
-/**
- * Class Data
- *
- * @category Magento
- * @package  Magezil_BoughtTogether
- * @author   Lucas Teixeira dos Santos Santana <santanaluc94@gmail.com>
- * @license  NO-LICENSE #
- * @link     http://github.com/santanaluc94
- */
-class Data extends AbstractHelper
+class Options
 {
     const MODULE_ENABLE = 'magezil_bought_together/general/enable';
     const BOUGHT_TOGETHER_LOGGED_IN = 'magezil_bought_together/general/user_logged';
@@ -26,8 +16,8 @@ class Data extends AbstractHelper
     const SHOW_COMPARE = 'magezil_bought_together/cards_configuration/show_compare';
     const SHOW_QTY_PRODUCTS = 'magezil_bought_together/cards_configuration/qty_products';
 
-    protected $scopeConfig;
-    protected $storeManager;
+    protected ScopeConfigInterface $scopeConfig;
+    protected StoreManagerInterface $storeManager;
 
     public function __construct(
         ScopeConfigInterface $scopeConfig,
@@ -37,9 +27,6 @@ class Data extends AbstractHelper
         $this->storeManager = $storeManager;
     }
 
-    /**
-     * Check module is enable
-     */
     public function isEnabled(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -48,9 +35,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Is bought together must be logged in
-     */
     public function isBoughtTogetherLoggedIn(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -59,9 +43,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Has Frequently Bought Together Title
-     */
     public function hasBoughtTogetherTitle(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -70,9 +51,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Get Frequently Bought Together Title
-     */
     public function getBoughtTogetherTitle(): string
     {
         return $this->scopeConfig->getValue(
@@ -81,9 +59,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Has quantity Products of Frequently Bought Together
-     */
     public function hasBoughtTogetherProductsQty(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -92,20 +67,14 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Get quantity Products of Frequently Bought Together
-     */
-    public function getBoughtTogetherProductsQty(): string
+    public function getBoughtTogetherProductsQty(): int
     {
-        return $this->scopeConfig->getValue(
+        return (int) $this->scopeConfig->getValue(
             self::BOUGHT_TOGETHER_QUANTITY,
             ScopeInterface::SCOPE_WEBSITE
         );
     }
 
-    /**
-     * Is show wishlist in bought together cards
-     */
     public function isShowWishlist(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -114,9 +83,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Is show compare in bought together cards
-     */
     public function isShowCompare(): bool
     {
         return $this->scopeConfig->isSetFlag(
@@ -125,9 +91,6 @@ class Data extends AbstractHelper
         );
     }
 
-    /**
-     * Is show quantity products in bought together cards
-     */
     public function isShowQtyProducts(): bool
     {
         return $this->scopeConfig->isSetFlag(
